Modular cloze editor.
