+++ components/survey-editor/components/ItemEditor/editor/components/content-editor/response-editors/cloze-editor.tsx
@@ -1,376 +1,60 @@
-import SortableItem from "@/components/survey-editor/components/general/SortableItem";
-import TabCard from "@/components/survey-editor/components/general/tab-card";
-import { ClozeItemType, ItemComponentRole } from "@/components/survey-editor/components/types";
-import { Badge } from "@/components/ui/badge";
-import { Collapsible, CollapsibleContent, CollapsibleTrigger } from "@/components/ui/collapsible";
-import { Input } from "@/components/ui/input";
-import { Label } from "@radix-ui/react-label";
-import { Separator } from "@radix-ui/react-separator";
-import { Binary, Calendar, ChevronDown, Clock, Cog, CornerDownLeft, FormInput, GripVertical, Heading, Languages, ToggleLeft } from "lucide-react";
-import { ItemComponent, ItemGroupComponent, SurveySingleItem } from "survey-engine/data_types";
-import DateInputContentConfig from "./date-input-content-config";
-import NumberInputContentConfig from "./number-input-content-config";
-import TextInputContentConfig from "./text-input-content-config";
-import { Button } from "@/components/ui/button";
-import SortableWrapper from "@/components/survey-editor/components/general/SortableWrapper";
-import AddDropdown from "@/components/survey-editor/components/general/add-dropdown";
-import React, { useContext } from "react";
-import { localisedObjectToMap } from "@/components/survey-editor/utils/localeUtils";
-import { generateLocStrings } from "case-editor-tools/surveys/utils/simple-generators";
-import { SurveyContext } from "@/components/survey-editor/surveyContext";
-import MarkdownContentEditor from "../markdown-content-editor";
+import { ItemComponentRole } from "@/components/survey-editor/components/types";
+import { ItemGroupComponent, SurveySingleItem } from "survey-engine/data_types";
+import React from "react";
+import ClozeContentConfig from "./cloze-content-config";
 
 interface ClozeEditorProps {
     surveyItem: SurveySingleItem;
     onUpdateSurveyItem: (item: SurveySingleItem) => void;
 }
 
-const getClozeItemType = (item: ItemComponent): ClozeItemType => {
-    const itemType = item.role as keyof typeof ClozeItemType;
-    return itemType as ClozeItemType;
-}
-
-export const TabWrapper = (props: { children: React.ReactNode }) => {
-    return (
-        <div className='p-4 ps-6 space-y-4 overflow-y-scroll'>
-            {props.children}
-        </div>
-    )
-}
-
-export const KeyAndType = (props: { compKey?: string, type: string }) => {
-    return <div className='text-xs font-semibold flex justify-between w-full'>
-        <Badge className='h-auto py-0'>
-            {props.compKey}
-        </Badge>
-        <span className='text-muted-foreground'>
-            {props.type}
-        </span>
-    </div>
-}
-
-export const ContentTabCollapsible = (props: { compKey?: string, type: string, children: React.ReactNode, defaultOpen: boolean }) => {
-    return <div className='space-y-4'>
-        <Collapsible defaultOpen={props.defaultOpen}
-            className='group'
-        >
-            <CollapsibleTrigger asChild>
-                <div className='flex w-full gap-2'>
-                    <KeyAndType compKey={props.compKey} type={props.type} />
-                    <span>
-                        <ChevronDown className="size-4 group-data-[state=open]:rotate-180 transition-transform duration-500" />
-                    </span>
-                </div>
-            </CollapsibleTrigger>
-            <CollapsibleContent className='pt-4'>
-                {props.children}
-            </CollapsibleContent>
-        </Collapsible>
-    </div>
-}
-
-export const ContentItem = (props: {
-    index: number, component: ItemComponent,
-    onUpdateComponent: (component: ItemComponent) => void,
-    onDeleteComponent: () => void,
-    existingKeys?: string[]
-
-}) => {
-    const { selectedLanguage } = useContext(SurveyContext);
-    const [currentKey, setCurrentKey] = React.useState(props.component.key || '');
-
-    const clozeItemType = getClozeItemType(props.component);
-
-    //console.log('optionType:', clozeItemType);
-
-    const renderContent = () => {
-        const currentContent = localisedObjectToMap(props.component.content).get(selectedLanguage) || '';
-        switch (clozeItemType) {
-            case ClozeItemType.SimpleText:
-                return <ContentTabCollapsible
-                    compKey={props.component.key}
-                    type='SIMPLE TEXT'
-                    defaultOpen={props.index > -1}>
-                    <div className='space-y-1.5'
-                        data-no-dnd="true">
-                        <Label htmlFor={props.component.key + '-text'} />
-                        <Input
-                            id={props.component.key + '-text'}
-                            placeholder='Text in selected language...'
-                            value={currentContent}
-                            onChange={(e) => {
-                                const updatedPart = { ...props.component };
-                                const updatedContent = localisedObjectToMap(updatedPart.content);
-                                updatedContent.set(selectedLanguage, e.target.value);
-                                updatedPart.content = generateLocStrings(updatedContent);
-                                props.onUpdateComponent(updatedPart);
-                            }}
-                        />
-                    </div>
-                </ContentTabCollapsible>;
-            case ClozeItemType.TextInput:
-                return <ContentTabCollapsible
-                    compKey={props.component.key}
-                    type='TEXT INPUT'
-                    defaultOpen={props.index > -1}
-                >
-                    <TextInputContentConfig
-                        component={props.component}
-                        onChange={props.onUpdateComponent}
-                        allowMultipleLines={true}
-                    />
-                </ContentTabCollapsible>;
-            case ClozeItemType.NumberInput:
-                return <ContentTabCollapsible
-                    compKey={props.component.key}
-                    type='NUMBER INPUT'
-                    defaultOpen={props.index > -1}
-                >
-                    <NumberInputContentConfig
-                        component={props.component}
-                        onChange={props.onUpdateComponent}
-                    />
-                </ContentTabCollapsible>;
-            case ClozeItemType.DateInput:
-                return <ContentTabCollapsible
-                    compKey={props.component.key}
-                    type='DATE INPUT'
-                    defaultOpen={props.index > -1}
-                >
-                    <DateInputContentConfig
-                        component={props.component}
-                        onChange={props.onUpdateComponent}
-                    />
-                </ContentTabCollapsible>;
-            case ClozeItemType.Markdown:
-                return <ContentTabCollapsible
-                    compKey={props.component.key}
-                    type='MARKDOWN'
-                    defaultOpen={props.index > -1}>
-                    <div className='space-y-1.5'
-                        data-no-dnd="true">
-                        <MarkdownContentEditor
-                            content={currentContent}
-                            onUpdateContent={(content) => {
-                                const updatedPart = { ...props.component };
-                                const updatedContent = localisedObjectToMap(updatedPart.content);
-                                updatedContent.set(selectedLanguage, content);
-                                updatedPart.content = generateLocStrings(updatedContent);
-                                props.onUpdateComponent(updatedPart);
-                            }}
-                        />
-                    </div>
-                </ContentTabCollapsible>;
-            case ClozeItemType.LineBreak:
-                return <div>Line break</div>;
-            default:
-                console.warn('Unknown / unimplemented item type:', clozeItemType);
-                return <div>Unknown / unimplemented role: {clozeItemType}</div>;
-        }
+const ClozeEditor: React.FC<ClozeEditorProps> = (props) => {
+    //const [draggedId, setDraggedId] = React.useState<string | null>(null);
+    //console.log('props.surveyItem:', props.surveyItem);
+    const rgIndex = props.surveyItem.components?.items.findIndex(comp => comp.role === ItemComponentRole.ResponseGroup);
+    if (rgIndex === undefined || rgIndex === -1) {
+        return <p>Response group not found</p>;
+    }
+    const responseGroup = props.surveyItem.components?.items[rgIndex] as ItemGroupComponent;
+    if (!responseGroup || !responseGroup.items) {
+        return <p>Response group not found</p>;
     }
 
-    return (<SortableItem
-        id={props.component.key || props.index.toString()}
-    >
-        <div className='relative'>
-            <div className='absolute left-0 top-1/2'>
-                <GripVertical className='size-4' />
-            </div>
-            <TabCard
-                tabs={[
-                    {
-                        label: 'Content',
-                        icon: <Languages className='me-1 size-3 text-muted-foreground' />,
-                        content: <TabWrapper>
-                            {renderContent()}
-                        </TabWrapper>
-                    },
-                    {
-                        label: 'Conditions',
-                        icon: <ToggleLeft className='me-1 size-3 text-muted-foreground' />,
-                        content: <TabWrapper>
-                            TODO: display and disabled condition editor - for section header, disabled part is not relevant
+    const clozeComponentIndex = responseGroup.items.findIndex(comp => comp.role === ItemComponentRole.Cloze);
+    if (clozeComponentIndex === undefined || clozeComponentIndex === -1) {
+        return <p>Cloze group not found</p>;
+    }
 
-                        </TabWrapper>
-                    },
-                    {
-                        label: 'Settings',
-                        icon: <Cog className='me-1 size-3 text-muted-foreground' />,
-                        content: <TabWrapper>
-                            <div className='space-y-1.5'
-                                data-no-dnd="true"
-                            >
-                                <Label
-                                    htmlFor='key'
-                                >
-                                    Key
-                                </Label>
-                                <div className='flex gap-2'>
-                                    <Input
-                                        id='key'
-                                        value={currentKey}
-                                        onChange={(e) => {
-                                            setCurrentKey(e.target.value);
-                                        }}
-                                    />
+    //const responseGroup = props.surveyItem.components?.items.find(item => item.role === ItemComponentRole.ResponseGroup) as ItemGroupComponent;
 
-                                    <Button
-                                        variant={'outline'}
-                                        disabled={currentKey === props.component.key}
-                                        onClick={() => {
-                                            setCurrentKey(props.component.key || '');
-                                        }}
-                                    >
-                                        Reset
-                                    </Button>
-                                    <Button
-                                        variant={'outline'}
-                                        disabled={currentKey === props.component.key || !currentKey || props.existingKeys?.includes(currentKey)}
-                                        onClick={() => {
-                                            props.onUpdateComponent({
-                                                ...props.component,
-                                                key: currentKey,
-                                            });
-                                        }}
-                                    >
-                                        Apply
-                                    </Button>
-                                </div>
-                                {(props.existingKeys?.includes(currentKey || '') && currentKey !== props.component.key) && <p className='text-danger-600 text-sm font-semibold'>Key already in use</p>}
+    /*     const updateSurveyItemWithNewItems = (newItems: ItemComponent[]) => {
+            responseGroup.items = responseGroup.items.map(comp => comp.role === ItemComponentRole.Cloze ? { ...comp, items: newItems } : comp);
+            const updatedItemGroupComponent = { ...props.surveyItem.components, items: props.surveyItem.components?.items.map(comp => comp.key === responseGroup.key ? responseGroup : comp) } as ItemGroupComponent;
+            props.onUpdateSurveyItem({
+                ...props.surveyItem,
+                components: updatedItemGroupComponent,
+            });
+        } */
 
-                            </div>
-                            <Separator />
-                            <Button
-                                data-no-dnd="true"
-                                variant={'outline'}
-                                className='hover:bg-danger-100'
-                                onClick={() => {
-                                    if (!confirm('Are you sure you want to delete this component?')) {
-                                        return;
-                                    }
-                                    props.onDeleteComponent();
-                                }}
-                            >
-                                Delete component
-                            </Button>
-                        </TabWrapper>
-                    },
-                ]}
-            />
-        </div>
-    </SortableItem>)
-}
-
-const ClozeEditor: React.FC<ClozeEditorProps> = (props) => {
-    const [draggedId, setDraggedId] = React.useState<string | null>(null);
-    //console.log('props.surveyItem:', props.surveyItem);
-
-
-    const responseGroup = props.surveyItem.components?.items.find(item => item.role === ItemComponentRole.ResponseGroup) as ItemGroupComponent;
-    //console.log('responseGroup:', responseGroup);
-    const clozeItems = (responseGroup.items.find(item => item.role === ItemComponentRole.Cloze) as ItemGroupComponent)?.items || [];
-    //console.log('clozeItems:', clozeItems);
-
-    const onAddItem = (keyOfSelectedItem: string) => {
-        const randomKey = Math.random().toString(36).substring(9);
-        //console.log('keyOfSelectedItem:', keyOfSelectedItem);
-        const itemType = keyOfSelectedItem as keyof typeof ClozeItemType;
-
-        // Create empty option of given type.
-        const newItem = {
-            key: randomKey,
-            role: itemType,
-        }
-
-        const newItems = [...clozeItems, newItem];
-        updateSurveyItemWithNewItems(newItems);
-    };
-
-    const draggedItem = clozeItems.find(comp => comp.key === draggedId);
-
-    const updateSurveyItemWithNewItems = (newItems: ItemComponent[]) => {
-        responseGroup.items = responseGroup.items.map(comp => comp.role === ItemComponentRole.Cloze ? { ...comp, items: newItems } : comp);
-        const updatedItemGroupComponent = { ...props.surveyItem.components, items: props.surveyItem.components?.items.map(comp => comp.key === responseGroup.key ? responseGroup : comp) } as ItemGroupComponent;
+    const onChange = (newComp: ItemGroupComponent) => {
+        const existingComponents = props.surveyItem.components?.items || [];
+        (existingComponents[rgIndex] as ItemGroupComponent).items[clozeComponentIndex] = newComp;
         props.onUpdateSurveyItem({
             ...props.surveyItem,
-            components: updatedItemGroupComponent,
+            components: {
+                role: props.surveyItem.components!.role,
+                ...props.surveyItem.components,
+                items: existingComponents,
+            }
         });
     }
 
-    const usedKeys = clozeItems.map(comp => comp.key || '');
 
     return (
-        <div className='mt-4'>
-            <p className='font-semibold mb-2'>Items: <span className='text-muted-foreground'>({clozeItems.length})</span></p>
-            <SortableWrapper
-                sortableID={'response-group-editor'}
-                direction='vertical'
-                items={clozeItems.map((component, index) => {
-                    return {
-                        id: component.key || index.toString(),
-                    }
-                })}
-                onDraggedIdChange={(id) => {
-                    setDraggedId(id);
-                }}
-                onReorder={(activeIndex, overIndex) => {
-                    const newItems = [...clozeItems];
-                    newItems.splice(overIndex, 0, newItems.splice(activeIndex, 1)[0]);
-
-                    updateSurveyItemWithNewItems(newItems);
-                }}
-                dragOverlayItem={(draggedId && draggedItem) ?
-                    <ContentItem
-                        index={-1}
-                        component={draggedItem}
-                        onDeleteComponent={() => { }}
-                        onUpdateComponent={() => { }}
-                    />
-                    : null}
-            >
-                <div className='overflow-y-scroll'>
-                    <ol className='flex flex-col gap-4 min-w-full'>
-                        {clozeItems.map((component, index) => {
-                            return <ContentItem
-                                key={component.key || index}
-                                index={index}
-                                component={component}
-                                existingKeys={usedKeys}
-                                onDeleteComponent={() => {
-                                    const newItems = clozeItems.filter(comp => comp.key !== component.key);
-                                    updateSurveyItemWithNewItems(newItems);
-                                }}
-                                onUpdateComponent={(updatedItem) => {
-                                    const newItems = clozeItems.map((comp => {
-                                        if (comp.key === component.key) {
-                                            return updatedItem;
-                                        }
-                                        return comp;
-                                    }))
-                                    updateSurveyItemWithNewItems(newItems);
-                                }}
-                            />
-                        })}
-
-                        <div className='flex justify-center w-full'>
-                            <AddDropdown
-                                options={[
-                                    { key: ClozeItemType.SimpleText, label: 'Text', icon: <Heading className='size-4 text-muted-foreground me-2' /> },
-                                    { key: ClozeItemType.Markdown, label: 'Markdown', icon: <Heading className='size-4 text-muted-foreground me-2' /> },
-                                    { key: ClozeItemType.TextInput, label: 'Text Input', icon: <FormInput className='size-4 text-muted-foreground me-2' /> },
-                                    { key: ClozeItemType.NumberInput, label: 'Number Input', icon: <Binary className='size-4 text-muted-foreground me-2' /> },
-                                    { key: ClozeItemType.TimeInput, label: 'Time Input', icon: <Clock className='size-4 text-muted-foreground me-2' /> },
-                                    { key: ClozeItemType.DateInput, label: 'Date Input', icon: <Calendar className='size-4 text-muted-foreground me-2' /> },
-                                    { key: ClozeItemType.LineBreak, label: 'Line Break', icon: <CornerDownLeft className='size-4 text-muted-foreground me-2' /> },
-                                ]}
-                                onAddItem={onAddItem}
-                            />
-                        </div>
-                    </ol>
-                </div>
-
-            </SortableWrapper>
-        </div>
+        <ClozeContentConfig
+            component={responseGroup.items[clozeComponentIndex] as ItemGroupComponent}
+            onChange={onChange}
+        />
     );
 }
